redirecciones

+++ repo_tpfinal/Vistas/Producto.aspx.cs
@@ -145,13 +145,13 @@ namespace Vistas
             {
                 InnerHTML = "";
                 DataTable infoCarrito = (DataTable)Session["Carrito"];
-                int TotalCarro = 0;
+                float TotalCarro = 0;
                 int CantProds = 0;
 
                 foreach (DataRow row in infoCarrito.Rows)
                 {
                     CantProds += int.Parse(row[1].ToString());
-                    TotalCarro += CantProds * int.Parse(row[2].ToString());
+                    TotalCarro += CantProds * float.Parse(row[2].ToString());
                 }
 
                 InnerHTML += "$" + TotalCarro + "(" + CantProds + ")";
